pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools"

[project]
name = "rag-chatbot-api"
version = "1.0.0"
description = "RAG Chatbot API for Physical AI & Humanoid Robotics textbook with FastAPI, Qdrant, and Google Gemini"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Physical AI & Humanoid Robotics Team"}
]
keywords = ["rag", "chatbot", "qdrant", "gemini", "fastapi", "docusaurus", "textbook"]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "qdrant-client>=1.7.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sse-starlette>=1.6.0",
    "cohere>=5.0.0",
    "asyncpg>=0.29.0",
    "pytest>=7.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0"
]

[project.scripts]
chat-api = "backend.src.main:app"

[tool.setuptools.packages.find]
where = ["backend/src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config"
]

[tool.coverage.run]
source = ["backend/src"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]